more code cleaning and refactoring
